+++ fjs/types/bigint/proof.f.ts
@@ -16,7 +16,7 @@ import {
     roundUp8
 } from './module.f.ts'
 import { assert, assertEq } from '../../asserts/module.f.mjs'
-import { min } from '../function/compare/module.f.ts'
+import { min } from '../function/compare/module.f.mjs'
 
 const oldLog2 = (v: bigint): bigint => {
     if (v <= 0n) { return -1n }
