+++ fjs/types/btree/find/module.f.ts
@@ -5,7 +5,7 @@
  */
 import type { Leaf1, Leaf2, Branch3, Branch5, TNode } from '../types/module.f.ts'
 import type { List } from '../../list/module.f.ts'
-import { index3, index5, type Compare } from '../../function/compare/module.f.ts'
+import { index3, index5, type Compare } from '../../function/compare/module.f.mjs'
 import type { KeyOf, Index, Tuple } from "../../array/module.f.mjs";
 
 export type FirstLeaf1<T> = readonly[Index<3>, Leaf1<T>]
