+++ fjs/types/btree/remove/module.f.ts
@@ -4,7 +4,7 @@
  * @module
  */
 import { collapseRoot, type Leaf1, type TNode, type Branch1, type Branch3, type Branch5, type Tree } from '../types/module.f.ts'
-import type { Compare } from '../../function/compare/module.f.ts'
+import type { Compare } from '../../function/compare/module.f.mjs'
 import { type Path, type PathItem, find } from '../find/module.f.ts'
 import { fold, concat, next } from '../../list/module.f.ts'
 import type { Tuple } from '../../array/module.f.mjs'
