+++ fjs/types/btree/set/module.f.ts
@@ -5,7 +5,7 @@
  */
 import { collapseRoot, type Branch1, type Branch3, type Branch5, type Branch7, type TNode, type Tree } from '../types/module.f.ts'
 import { find, type First, type PathItem, type Result } from '../find/module.f.ts'
-import type { Compare } from '../../function/compare/module.f.ts'
+import type { Compare } from '../../function/compare/module.f.mjs'
 import { fold } from '../../list/module.f.ts'
 
 type Branch1To3<T> = Branch1<T> | Branch3<T>
