+++ fjs/types/number/module.f.ts
@@ -6,7 +6,7 @@
  */
 import { reduce, type List } from '../list/module.f.ts'
 import { addition } from '../function/operator/module.f.ts'
-import { type Sign, cmp as uCmp, min as uMin, max as uMax } from '../function/compare/module.f.ts'
+import { type Sign, cmp as uCmp, min as uMin, max as uMax } from '../function/compare/module.f.mjs'
 import { fold } from '../../common/monoid/module.f.ts'
 
 export const sum: (input: List<number>) => number
