+++ fjs/types/ordered_map/module.f.ts
@@ -8,7 +8,7 @@ import { value, find } from '../btree/find/module.f.ts'
 import { set } from '../btree/set/module.f.ts'
 import { remove as btreeRemove } from '../btree/remove/module.f.ts'
 import { values } from '../btree/module.f.ts'
-import type { Sign } from '../function/compare/module.f.ts'
+import type { Sign } from '../function/compare/module.f.mjs'
 import { cmp } from '../string/module.f.ts'
 import { fold, type List } from '../list/module.f.ts'
 import type { Reduce } from '../function/operator/module.f.ts'
