+++ fjs/types/sorted_list/module.f.ts
@@ -3,7 +3,7 @@
  *
  * @module
  */
-import { type Sign, type Cmp, bsearch } from '../function/compare/module.f.ts'
+import { type Sign, type Cmp, bsearch } from '../function/compare/module.f.mjs'
 import { type List, next } from '../list/module.f.ts'
 import type { Nullable } from '../nullable/module.f.mjs'
 import { identity } from '../function/module.f.mjs'
