+++ fjs/types/string/module.f.ts
@@ -17,7 +17,7 @@
  */
 import { type List, reduce as listReduce, repeat as listRepeat } from '../list/module.f.ts'
 import { compose } from '../function/module.f.mjs'
-import { type Sign, cmp as uCmp } from '../function/compare/module.f.ts'
+import { type Sign, cmp as uCmp } from '../function/compare/module.f.mjs'
 import { join as joinOp, type Reduce } from '../function/operator/module.f.ts'
 import { fold } from '../../common/monoid/module.f.ts'
